+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Sync to upstream Luau 0.670
 - Linux runners for GitHub actions are bumped from the deprecated `ubuntu-20.04` to `ubuntu-22.04`. This may mean that release artifacts no longer work on `ubuntu-20.04`.
 - Improved Studio plugin error message when attempting to connect to a server but it is not online ([#982](https://github.com/JohnnyMorganz/luau-lsp/issues/982))
+- VSCode extension will now strip the prefix `FFlag` / `DFFlag` / `FInt` / `DFInt` if they were included in the overrides name ([#981](https://github.com/JohnnyMorganz/luau-lsp/issues/981))
 
 ### Fixed